backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
langchain-google-genai==0.0.6
sqlalchemy==2.0.23
plotly==5.18.0
google-generativeai==0.3.2
pandas==2.1.3
numpy==1.26.2
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
sentence-transformers==2.2.2 